cabinet/requirements.txt
email_validator>=2.2.0,<3.0.0
Flask-Login>=0.6.3,<1.0.0
Flask-Migrate>=4.0.7,<5.0.0
Flask-Session>=0.8.0,<1.0.0
Flask-SQLAlchemy>=3.1.1,<4.0.0
Flask-WTF>=1.2.2,<2.0.0
Flask>=3.1.0,<4.0.0
psycopg2-binary>=2.9.10,<3.0.0
redis>=5.2.1,<6.0.0
requests>=2.32.3,<3.0.0
splunk-opentelemetry[all]>=1.21.0,<2.0.0
SQLAlchemy>=2.0.36,<3.0.0
waitress>=3.0.2,<4.0.0